+++ configure.py
@@ -1061,6 +1061,7 @@ scylla_core = (['message/messaging_service.cc',
                 'cql3/statements/prune_materialized_view_statement.cc',
                 'cql3/statements/batch_statement.cc',
                 'cql3/statements/select_statement.cc',
+                'cql3/statements/external_search/vector_indexed_table_select_statement.cc',
                 'cql3/statements/use_statement.cc',
                 'cql3/statements/index_prop_defs.cc',
                 'cql3/statements/index_target.cc',
