+++ cql3/statements/select_statement.hh
@@ -21,8 +21,6 @@
 #include "exceptions/coordinator_result.hh"
 #include "locator/host_id.hh"
 #include "service/cas_shard.hh"
-#include "vector_search/vector_store_client.hh"
-#include "vector_search/filter.hh"
 
 namespace service {
     class client_state;
@@ -360,49 +358,6 @@ private:
             service::query_state& state, const query_options& options) const override;
 };
 
-
-class vector_indexed_table_select_statement : public select_statement {
-    secondary_index::index _index;
-    prepared_ann_ordering_type _prepared_ann_ordering;
-    vector_search::prepared_filter _prepared_filter;
-    mutable gc_clock::time_point _query_start_time_point;
-
-public:
-    static constexpr size_t max_ann_query_limit = 1000;
-
-    static ::shared_ptr<cql3::statements::select_statement> prepare(data_dictionary::database db, schema_ptr schema, uint32_t bound_terms,
-            lw_shared_ptr<const parameters> parameters, ::shared_ptr<selection::selection> selection,
-            ::shared_ptr<const restrictions::statement_restrictions> restrictions, ::shared_ptr<std::vector<size_t>> group_by_cell_indices, bool is_reversed,
-            ordering_comparator_type ordering_comparator, prepared_ann_ordering_type prepared_ann_ordering, std::optional<expr::expression> limit,
-            std::optional<expr::expression> per_partition_limit, cql_stats& stats, const secondary_index::index& index, std::unique_ptr<cql3::attributes> attrs);
-
-    vector_indexed_table_select_statement(schema_ptr schema, uint32_t bound_terms, lw_shared_ptr<const parameters> parameters,
-            ::shared_ptr<selection::selection> selection, ::shared_ptr<const restrictions::statement_restrictions> restrictions,
-            ::shared_ptr<std::vector<size_t>> group_by_cell_indices, bool is_reversed, ordering_comparator_type ordering_comparator,
-            prepared_ann_ordering_type prepared_ann_ordering, std::optional<expr::expression> limit, std::optional<expr::expression> per_partition_limit,
-            cql_stats& stats, const secondary_index::index& index, vector_search::prepared_filter prepared_filter, std::unique_ptr<cql3::attributes> attrs);
-
-private:
-    future<::shared_ptr<cql_transport::messages::result_message>> do_execute(
-            query_processor& qp, service::query_state& state, const query_options& options) const override;
-
-    void update_stats() const;
-
-    lw_shared_ptr<query::read_command> prepare_command_for_base_query(query_processor& qp, service::query_state& state, const query_options& options, uint64_t fetch_limit) const;
-
-    std::vector<float> get_ann_ordering_vector(const query_options& options) const;
-
-    future<::shared_ptr<cql_transport::messages::result_message>> query_base_table(query_processor& qp, service::query_state& state,
-            const query_options& options, const std::vector<vector_search::primary_key>& pkeys, lowres_clock::time_point timeout) const;
-
-    future<coordinator_result<foreign_ptr<lw_shared_ptr<query::result>>>> query_base_table(query_processor& qp, service::query_state& state,
-            const query_options& options, lw_shared_ptr<query::read_command> command, lowres_clock::time_point timeout,
-            const std::vector<vector_search::primary_key>& pkeys) const;
-
-    future<coordinator_result<foreign_ptr<lw_shared_ptr<query::result>>>> query_base_table(query_processor& qp, service::query_state& state,
-            const query_options& options, lw_shared_ptr<query::read_command> command, lowres_clock::time_point timeout,
-            std::vector<dht::partition_range> partition_ranges) const;
-};
 
 }
 }
